guid request and mail sent

+++ src/main/java/org/example/smallworld_backend/dto/GuidDTO.java
@@ -19,12 +19,12 @@ public class GuidDTO {
     private String address;
     private String about_me;
     private List<String> languages;
-    private double years_experience;
+    private int years_experience;
     private List<String> certifications;
     private String price;
     private String availability;
 
-    public GuidDTO(String profile_image, String name, String email, String phone_number, String address, String about_me, List<String> languages, double years_experience, List<String> certifications, String price, String availability) {
+    public GuidDTO(String profile_image, String name, String email, String phone_number, String address, String about_me, List<String> languages, int years_experience, List<String> certifications, String price, String availability) {
         this.profile_image = profile_image;
         this.name = name;
         this.email = email;
